Improve UPnP IGD selection.
Previously, every valid IGD that sent us a notify bulldozed the last. This meant that if you have multiple such devices on your network (of which only one is the actual router), the UPnP stuff would not necessarily select the correct one (and also race condition... yay).

This is now fixed. Allow valid UPnP devices are queued up and a pick is only made at the end of the UPnP discovery timeout. This selection checks whether the device actually has a proper external address to weed out invalid ones.

(My upstairs wifi AP implements UPnP because I think it's designed to be usable as a wifi-only router. Lidgren was preferring it over my real router)