Warning/error logs don't allocate if disabled.
This should completely eliminate all performance overhead from the logging when it's disabled. This proves relevant during DDoS attacks, as the logging is actually a serious amplifier for even low-bandwidth UDP spam.

This is done by moving it to an interpolated string handler, so we can skip running string formatting and such when the log channel is disabled.